Fix failing unit tests for DigitalOcean App Platform Service and GCP resources
- Add missing CEL validation rules to DigitalOcean App Platform Service spec:
  - source_required: Ensures either git_source or image_source is specified
  - autoscale_config: Validates min/max instance counts when autoscale is enabled

- Fix duplicate RunSpecs issue in GCP and Snowflake tests:
  - Remove redundant api_test.go files that were causing 'RunSpecs called twice' error
  - Consolidate tests into spec_test.go files for 13 cloud resources
  - Update BUILD.bazel files to reflect removed test files

All tests now pass successfully.
