+++ MODULE.bazel
@@ -13,9 +13,15 @@ go_sdk.download(version = "1.23.4")
 # Read Go deps from this repo's go.mod (bzlmod-friendly)
 go_deps = use_extension("@gazelle//:extensions.bzl", "go_deps")
 go_deps.from_file(go_mod = "//:go.mod")
+go_deps.module(
+    path = "buf.build/go/protovalidate",
+    sum = "h1:kr/rC/no+DtRyYX+8KXLDxNnI1rINz0imk5K44ZpZ3A=",
+    version = "v0.14.0",
+)
 use_repo(
     go_deps,
     "build_buf_gen_go_bufbuild_protovalidate_protocolbuffers_go",
+    "build_buf_go_protovalidate",
     "com_github_blang_semver",
     "com_github_google_uuid",
     "com_github_hashicorp_hcl_v2",
