+++ apis/project/planton/provider/aws/awsstaticwebsite/v1/spec_test.go
