+++ apis/project/planton/provider/digitalocean/digitaloceanvpc/v1/spec_test.go
@@ -6,8 +6,8 @@ import (
 	"buf.build/go/protovalidate"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/project-planton/project-planton/apis/project/planton/shared"
 	"github.com/project-planton/project-planton/apis/project/planton/provider/digitalocean"
+	"github.com/project-planton/project-planton/apis/project/planton/shared"
 )
 
 func TestDigitalOceanVpcSpec(t *testing.T) {
@@ -22,13 +22,13 @@ var _ = Describe("DigitalOceanVpcSpec Custom Validation Tests", func() {
 
 			It("should not return a validation error for minimal valid fields", func() {
 				input := &DigitalOceanVpc{
-					ApiVersion: "digitalocean.project-planton.org/v1",
+					ApiVersion: "digital-ocean.project-planton.org/v1",
 					Kind:       "DigitalOceanVpc",
 					Metadata: &shared.ApiResourceMetadata{
 						Name: "test-vpc",
 					},
 					Spec: &DigitalOceanVpcSpec{
-						Region:       digitalocean.DigitalOceanRegion_nyc3,
+						Region:      digitalocean.DigitalOceanRegion_nyc3,
 						IpRangeCidr: "10.10.0.0/16",
 					},
 				}
