+++ apis/project/planton/provider/gcp/gcpartifactregistryrepo/v1/BUILD.bazel
@@ -22,10 +22,7 @@ go_library(
 
 go_test(
     name = "gcpartifactregistryrepo_test",
-    srcs = [
-        "api_test.go",
-        "spec_test.go",
-    ],
+    srcs = ["spec_test.go"],
     embed = [":gcpartifactregistryrepo"],
     deps = [
         "//apis/project/planton/shared",
