+++ apis/project/planton/provider/gcp/gcpdnszone/v1/BUILD.bazel
@@ -23,14 +23,10 @@ go_library(
 
 go_test(
     name = "gcpdnszone_test",
-    srcs = [
-        "api_test.go",
-        "spec_test.go",
-    ],
+    srcs = ["spec_test.go"],
     embed = [":gcpdnszone"],
     deps = [
         "//apis/project/planton/shared",
-        "//apis/project/planton/shared/networking/enums/dnsrecordtype",
         "@build_buf_go_protovalidate//:protovalidate",
         "@com_github_onsi_ginkgo_v2//:ginkgo",
         "@com_github_onsi_gomega//:gomega",
