+++ apis/project/planton/provider/gcp/gcpgkeaddonbundle/v1/BUILD.bazel
@@ -22,10 +22,7 @@ go_library(
 
 go_test(
     name = "gcpgkeaddonbundle_test",
-    srcs = [
-        "api_test.go",
-        "spec_test.go",
-    ],
+    srcs = ["spec_test.go"],
     embed = [":gcpgkeaddonbundle"],
     deps = [
         "//apis/project/planton/shared",
