+++ apis/project/planton/provider/gcp/gcpgkecluster/v1/BUILD.bazel
@@ -23,10 +23,7 @@ go_library(
 
 go_test(
     name = "gcpgkecluster_test",
-    srcs = [
-        "api_test.go",
-        "spec_test.go",
-    ],
+    srcs = ["spec_test.go"],
     embed = [":gcpgkecluster"],
     deps = [
         "//apis/project/planton/shared",
