+++ apis/project/planton/provider/gcp/gcpsecretsmanager/v1/BUILD.bazel
@@ -21,10 +21,7 @@ go_library(
 
 go_test(
     name = "gcpsecretsmanager_test",
-    srcs = [
-        "api_test.go",
-        "spec_test.go",
-    ],
+    srcs = ["spec_test.go"],
     embed = [":gcpsecretsmanager"],
     deps = [
         "//apis/project/planton/shared",
