+++ apis/project/planton/provider/gcp/gcpserviceaccount/v1/BUILD.bazel
@@ -22,10 +22,7 @@ go_library(
 
 go_test(
     name = "gcpserviceaccount_test",
-    srcs = [
-        "api_test.go",
-        "spec_test.go",
-    ],
+    srcs = ["spec_test.go"],
     embed = [":gcpserviceaccount"],
     deps = [
         "//apis/project/planton/shared",
